Fix ButtonTrait
