+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphRelatedContent.php
@@ -5,7 +5,7 @@ namespace Drupal\server_general\Plugin\EntityViewBuilder;
 use Drupal\paragraphs\ParagraphInterface;
 use Drupal\pluggable_entity_view_builder\EntityViewBuilderPluginAbstract;
 use Drupal\server_general\ProcessedTextBuilderTrait;
-use Drupal\server_general\ThemeTrait\ButtonTrait;
+use Drupal\server_general\ButtonTrait;
 use Drupal\server_general\ThemeTrait\CarouselTrait;
 
 /**
