+++ web/modules/custom/server_general/src/ThemeTrait/ButtonTrait.php
@@ -51,58 +51,4 @@ trait ButtonTrait {
       '#open_new_tab' => $open_new_tab,
     ];
   }
-
-  /**
-   * Get button by link field.
-   *
-   * @param \Drupal\Core\Entity\ContentEntityInterface $entity
-   *   The entity.
-   * @param string $field_name
-   *   The machine name of the field holding the link. Defaults to `field_link`.
-   *
-   * @return array
-   *   Render array.
-   */
-  protected function buildLinkButton(ContentEntityInterface $entity, string $field_name = 'field_link'): array {
-    if ($entity->get($field_name)->isEmpty()) {
-      return [];
-    }
-
-    $value = $this->getLinkFieldValue($entity, $field_name);
-    if (empty($value)) {
-      return [];
-    }
-
-    // If title is empty, show the URL itself.
-    $title = $value['title'] ?? $value['url']->toString();
-    return $this->buildButton($title, $value['url']);
-  }
-
-  /**
-   * Build a download file button from field.
-   *
-   * @param \Drupal\Core\Entity\ContentEntityInterface $entity
-   *   The entity.
-   * @param string $field_name
-   *   The machine name of the field holding the link. Defaults to `field_link`.
-   *
-   * @return array
-   *   Render array.
-   */
-  protected function buildFileButton(ContentEntityInterface $entity, string $field_name = 'field_file'): array {
-    /** @var \Drupal\file\FileInterface $file */
-    $file = $this->getReferencedEntityFromField($entity, $field_name);
-    if (!$file instanceof FileInterface) {
-      return [];
-    }
-
-    $value = $this->getLinkFieldValue($entity, $field_name);
-    if (empty($value)) {
-      return [];
-    }
-
-    $title = $value['title'] ?? $this->t('Download');
-    return $this->buildButton($title, $value['url']);
-  }
-
 }
